+++ lib/generators/compose/compose_pag_service.dart
@@ -406,7 +406,9 @@ class ComposePagService {
         error = classInfo;
       } else if (path.endsWith('_empty_feedback.dart') && empty == null) {
         empty = classInfo;
-      } else if (path.endsWith('_skeleton.dart') && skeleton == null) {
+      } else if (path.endsWith('_skeleton.dart') &&
+          skeleton == null &&
+          !path.contains('parts')) {
         skeleton = classInfo;
       }
     }